# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CategoriesListResponse:
    properties:
      categories:
        items:
          $ref: '#/definitions/handlers.CategoryResponse'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
  handlers.CategoryResponse:
    properties:
      id:
        type: string
      name:
        type: string
      product_count:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
...
  handlers.CreateCategoryRequest:
    properties:
      name:
        example: Electronics
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  handlers.StandardErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.UpdateCategoryRequest:
    properties:
      name:
        example: Electronics
        maxLength: 100
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  models.Category:
    properties:
      id:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
    type: object
  models.Comment:
    properties:
      body:
        type: string
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      created_at:
        type: string
      downvotes:
        type: integer
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      sentiment_label:
        type: string
      sentiment_score:
        type: number
      upvotes:
        type: integer
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.CommentVote:
    properties:
      comment:
        $ref: '#/definitions/models.Comment'
      comment_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
      vote_type:
        type: string
    type: object
  models.Favorite:
    properties:
      favorited_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.HttpRequestLog:
    properties:
      attack_score:
        type: number
      duration_ms:
        type: integer
      id:
        type: string
      ip_address:
        type: string
      method:
        type: string
      path:
        type: string
      referrer:
        type: string
      status_code:
        type: integer
      suspected_attack_type:
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: string
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      created_by:
        type: string
      creator:
        $ref: '#/definitions/models.User'
      curated_price:
        type: number
      curated_tags:
        items:
          type: string
        type: array
      description:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      feature_vector:
        $ref: '#/definitions/models.ProductFeatureVector'
      id:
        type: string
      name:
        type: string
      price:
        type: number
      product_suggestions:
        items:
          $ref: '#/definitions/models.ProductSuggestion'
        type: array
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      similarity_data_as_product:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      similarity_data_as_similar:
        items:
          $ref: '#/definitions/models.ProductSimilarityData'
        type: array
      tags:
        items:
          type: string
        type: array
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
  models.ProductFeatureVector:
    properties:
      embedding:
        items:
          type: number
        type: array
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      updated_at:
        type: string
    type: object
  models.ProductSimilarityData:
    properties:
      based_on:
        type: string
      created_at:
        type: string
      id:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      similar_product:
        $ref: '#/definitions/models.Product'
      similar_product_id:
        type: string
      similarity_score:
        type: number
    type: object
  models.ProductSuggestion:
    properties:
      generated_at:
        type: string
      id:
        type: string
      model_version:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      reason:
        type: string
      suggested_price_max:
        type: number
      suggested_price_min:
        type: number
      suggested_tags:
        items:
          type: string
        type: array
    type: object
  models.Recommendation:
    properties:
      age_based:
        type: boolean
      based_on_comments:
        type: boolean
      based_on_favorites:
        type: boolean
      created_at:
        type: string
      id:
        type: string
      model_version:
        type: string
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      reason:
        type: string
      region_based:
        type: boolean
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
  models.User:
    properties:
      birth_year:
        type: integer
      comment_votes:
        items:
          $ref: '#/definitions/models.CommentVote'
        type: array
      comments:
        items:
          $ref: '#/definitions/models.Comment'
        type: array
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      http_request_logs:
        items:
          $ref: '#/definitions/models.HttpRequestLog'
        type: array
      id:
        type: string
      is_admin:
        type: boolean
      name:
        type: string
      recommendations:
        items:
          $ref: '#/definitions/models.Recommendation'
        type: array
      region:
        type: string
      user_events:
        items:
          $ref: '#/definitions/models.UserEvent'
        type: array
    type: object
  models.UserEvent:
    properties:
      event_type:
        type: string
      id:
        type: string
      metadata:
        additionalProperties: true
        type: object
      product:
        $ref: '#/definitions/models.Product'
      product_id:
        type: string
      timestamp:
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@bachelor-ecommerce.com
    name: API Support
    url: http://www.swagger.io/support
  description: ML-Powered E-commerce Platform API with advanced recommendation system,
    intelligent search, and comprehensive analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Bachelor E-commerce API
  version: "1.0"
paths:
...
  /categories:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of all categories with product counts
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Categories retrieved successfully
          schema:
            $ref: '#/definitions/handlers.CategoriesListResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      summary: Get all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new product category
      parameters:
      - description: Category creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Category created successfully
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "409":
          description: Category with this name already exists
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new category
      tags:
      - Categories
  /categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a category (only if it has no products)
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Category deleted successfully
        "400":
          description: Invalid category ID or category has products
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete category
      tags:
      - Categories
    get:
      consumes:
      - application/json
      description: Retrieve a specific category with its product count
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Category retrieved successfully
          schema:
            $ref: '#/definitions/handlers.CategoryResponse'
        "400":
          description: Invalid category ID
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      summary: Get category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update an existing category's information
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Category updated successfully
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "401":
          description: User not authenticated
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "404":
          description: Category not found
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "409":
          description: Category with this name already exists
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.StandardErrorResponse'
      security:
      - BearerAuth: []
      summary: Update category
      tags:
      - Categories
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...